pyproject.toml
[project]
name = "self-healing-time-parser"
version = "0.1.0"
description = "Self-healing time parser system using LLM-based coding agent"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "python-dateutil>=2.8.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]

[tool.setuptools]
packages = ["time_parser", "coding_agent", "utils"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["time_parser/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

